feat[BREAKING]: result classmethods ok & err are now as_ok & as_err
The Result class contains two classmethods Result.Ok and Result.Err
as a quick way of making the Ok and Err variants. To improve clarity,
the constructors were renamed to Result.as_Ok and Result.as_Err
to improve the clarity. This change was propagated in the docstrings
and pytests.