refactor: convert empty_init constructor to private
The empty_init constructor is meant for making a result object
in an uninitialized state. This feature should only be used within the class
or by advanced users, so it was changed to _empty_init (private flag).

This change also removed the check in all methods for the Result
being initialized empty.