feat[BREAKING]: merge _ok and _err attributes to _val
The main attributes for storing the Ok result, self._Ok, and
the error result, self._Err, were merged into a single attribute
self._val, which either holds the value or a ResultErr object.

This simplifies the code and relies on the self._success attribute
to determine if _val is the Ok or Err variant.

The Result object also now has the limitation that you can NOT
wrap a ResultErr object in a Ok variant result.