fix: result.Ok to keep debugpy from converting Ok to Err
This is a strange bug in debugpy that randomly changes
an Ok variant to an Err variant. What happens is in the
background the python debugger evaluates all class properties.
But evaluating result.Err for an Ok(value) converts it to
an Err(e) and returns itself (since it resulted in an error).

This leads to random changing of Ok(value) variables by
the debugger, that would then work perfectly fine under
normal execution. This fix for changes both the Ok and
Err properties to return a new instance of an Err(e)
if they result in an error and DOES NOT alter self.