refactor[BREAKING]: resulterr remove error codes and groups
The ResultErr class defined an attribute `code` that stored an
int error codes and maintained a dictionary of error codes to
error definitions. This feature made the rest of the ResultContainer
overly complex for minimal added benefit, so it was removed.

This also removed the class based error group dictionaries.
A backup of the original python, with error codes, will be
archived on the error_codes branch.